update fakes ns, moved helpers to class
