+++ TusLibros.UnitTests/MerchantFakes/MerchantSpy.cs
@@ -1,6 +1,6 @@
 using TusLibros;
 
-namespace TusLibros.MerchantFakes
+namespace TusLibros.UnitTests.MerchantFakes
 {
     // Un Spy o Espía es un Fake que graba la información que se le envía.
     public class MerchantSpy : IMerchantAdapter
